shortcut-bot-with-gif/requirements.txt
Jinja2
pandas
python-dotenv
python-twitter
